+++ Changelog.md
@@ -33,6 +33,7 @@
 - allow scaling of loot and experience when the mob dies
   - option `scale_loot` to adjust whether or not the amount of dropped loot should be adjusted according to the scaling
   - option `scale_xp` to adjust whether or not the amount of dropped xp points should be adjusted according to the scaling
+- rename command `/mobScalings` to `/mobscalings`
 
 ### Version 1.20.1 - 1.3